+++ private/DeploymentState.class.ps1
@@ -1,5 +1,5 @@
 class SynapseDeploymentState {
-    [datetime] $LastUpdate = [System.DateTime]::UtcNow
+    [datetime] $LastUpdate
     [hashtable] $Deployed = @{}
     [string] $synapsetoolsVer = ''
     [string] $Algorithm = 'MD5'
@@ -34,7 +34,7 @@ class SynapseDeploymentState {
             $this.Deployed = Remove-ItemFromCollection -col $this.Deployed -item $_
             Write-Verbose "[DELETED] hash for $_"
         }
-        $this.LastUpdate
+        $this.LastUpdate = [System.DateTime]::UtcNow
         $this.StorageAccountName
         return $cnt;
     }
@@ -56,11 +56,27 @@ function Get-StateFromService {
 
         try {
             $StorageContext = New-AzStorageContext -StorageAccountName $StorageAccountName -ErrorAction Stop
-            $StorageContainer = Get-AzStorageContainer -Name 'azure-synapse-tools' -Context $StorageContext -ErrorAction Stop
+            $StorageContainer = Get-AzStorageContainer -Name azure-synapse-tools -Context $StorageContext -ErrorAction Stop
             $DeploymentStateFile = $StorageContainer.CloudBlobContainer.GetBlockBlobReference("$($targetSynapse.name)_deployment_state.json")
-            $res = $DeploymentStateFile.DownloadText()
-            if ($res) {
-                $res = $res | ConvertFrom-Json
+            Try {
+                $res = $DeploymentStateFile.DownloadText()
+                if ($res) {
+                    Try {
+                        $res = $res |ConvertFrom-Json -ErrorAction Stop
+                    }
+                    Catch {
+                        throw $_.Exception
+                    }
+                }
+            }
+            Catch {
+                Try {
+                    $DeploymentStateFile.UploadText('{"Deployed": {}}')
+                    Write-Host "Created placeholder $($targetSynapse.name)_deployment_state.json file"
+                }
+                Catch {
+                    throw $_.Exception
+                }
             }
         }
         catch {
@@ -90,7 +106,7 @@ function Set-StateFromService {
 
     try {
         $StorageContext = New-AzStorageContext -StorageAccountName $StorageAccountName -ErrorAction Stop
-        $StorageContainer = Get-AzStorageContainer -Name 'azure-synapse-tools' -Context $StorageContext -ErrorAction Stop
+        $StorageContainer = Get-AzStorageContainer -Name azure-synapse-tools -Context $StorageContext -ErrorAction Stop
         $DeploymentStateFile = $StorageContainer.CloudBlobContainer.GetBlockBlobReference("$($targetSynapse.name)_deployment_state.json")
         $DeploymentStateFile.UploadText($content)
         Write-Output "Successfully updated $($targetSynapse.name)_deployment_state.json"